Extract Keybinds to their own file
